chore: restructure project for better code organization
